requirements.txt
a2wsgi==1.10.10
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosmtplib==5.0.0
aiosqlite==0.21.0
alembic==1.17.1
altair==4.2.2
annotated-types==0.7.0
anyio==4.11.0
apache-airflow==3.1.2
apache-airflow-core==3.1.2
apache-airflow-providers-cncf-kubernetes==10.9.0
apache-airflow-providers-common-compat==1.8.0
apache-airflow-providers-common-io==1.6.4
apache-airflow-providers-common-sql==1.28.2
apache-airflow-providers-fab==3.0.1
apache-airflow-providers-smtp==2.3.1
apache-airflow-providers-standard==1.9.1
apache-airflow-task-sdk==1.1.2
apispec==6.8.4
argcomplete==3.6.3
asgiref==3.10.0
attrs==25.4.0
babel==2.17.0
blinker==1.9.0
cachelib==0.13.0
cachetools==6.2.2
cadwyn==5.4.5
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
clickclick==20.10.2
colorama==0.4.6
colorlog==6.10.1
connexion==2.14.2
cron_descriptor==2.0.6
croniter==6.0.0
cryptography==46.0.3
Deprecated==1.3.1
dill==0.4.0
dnspython==2.8.0
durationpy==0.10
email-validator==2.3.0
entrypoints==0.4
Faker==37.12.0
fastapi==0.117.1
fastapi-cli==0.0.16
Flask==2.2.5
Flask-AppBuilder==5.0.0
flask-babel==4.0.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Session==0.8.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.2
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.43.0
googleapis-common-protos==1.72.0
great_expectations==1.9.0
greenback==1.2.1
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==33.1.0
kubernetes_asyncio==33.3.0
lazy-object-proxy==1.12.0
libcst==1.8.6
limits==5.6.0
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
marshmallow-sqlalchemy==1.4.2
mdurl==0.1.2
methodtools==0.4.7
mistune==3.1.4
more-itertools==10.8.0
msgpack==1.1.2
msgspec==0.19.0
multidict==6.7.0
narwhals==2.11.0
natsort==8.4.0
numpy==2.3.4
oauthlib==3.3.1
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
patsy==1.0.2
pendulum==3.1.0
pillow==12.0.0
plotly==6.4.0
plotly-express==0.4.1
pluggy==1.6.0
prison==0.2.1
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.3
psycopg2-binary==2.9.11
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pygtrie==2.5.0
PyJWT==2.10.1
pyparsing==3.2.5
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rich==13.9.4
rich-argparse==1.7.2
rich-toolkit==0.15.1
rpds-py==0.28.0
rsa==4.9.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
scipy==1.16.3
setproctitle==1.3.7
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.42.0
sqlparse==0.5.3
starlette==0.48.0
statsmodels==0.14.5
streamlit==1.51.0
streamlit-aggrid==1.2.0.post1
structlog==25.5.0
svcs==25.1.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.2.0
text-unidecode==1.3
toml==0.10.2
toolz==1.1.0
tornado==6.5.2
tqdm==4.67.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uc-micro-py==1.0.3
universal_pathlib==0.2.6
urllib3==2.5.0
uuid6==2025.0.1
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==2.2.3
wirerope==1.0.0
wrapt==2.0.1
WTForms==3.2.1
yarl==1.22.0
zipp==3.23.0
